requirements.txt
pandas
playwright
feedgen
requests
feedparser
dateparser
pytz
python-dateutil
bs4
